Code changes required by updated parser. (Minimal changes required, yay for a mostly-stable API \o)
Change how parser callbacks are registered. Rather than needing to
register each one manually, we now register for every callback that
we implement the interface for.

Some changes to parser handling so as not to assume that we always have
an IRCParser. (We do, but lets not assume it!)

Fix not actually sending PINGs before timing out a client anymore

IRCServerType now correctly asks the IRCConnectionHandler to close
it's connections when it should.