Added some basic output filtering support (Issue #40)
Commands passed to the bot are split by "|" which denotes the start of a filter.

If a | is needed at the start of a word, then it should be escaped with a || - this is only needed once, and at the start of a word. so:
 - "|foo" is treated as the filter "foo"
 - "||foo" is treated as the string "|foo"
 - "|||foo" is treated as the string "||foo"

Current valid filters are "include" and "exclude", which compare the command output against the provided regex and either include it in or exclude it from the output sent to the user.

If an invalid filter is passed, or the filter fails to run then the command output will be returned entirely unfiltered.