BATCH specification was changed, again...
Really tempted to just remove support for this at the moment, as a result of the IRCv3 working group doign a really horrible job at specifiying it.

https://github.com/atheme/ircv3-specifications/commits/master/extensions/batch-3.2

It's changed 3 times so far in non-backwards compatible ways.
- Firstly optional reference
- Then they made the reference mandatory (Our initial implementation happened after this)
- Then they added a parameter to specify in advance how many lines the batch had and added a tag to the batched lines
- Then they removed the number of lines and added a batch type (of which only "NETSPLIT" and "NETJOIN" are valid)

For now I'm using a type of "generic" because I cba with a vendor-specific tag, and this will do for now as I imagine nothing much actually supports this...