Add ability to figure out Client Types (Part of Issue #51).
This will attempt to parse the VERSION reply to figure it out, but not all
clients actually send a version, so allow clients to force a type by
passing an extra paramerter after the subclient in the username field:

eg: dataforce+subclient+clienttype:password

subclient can be blank (eg dataforce++clienttype) if needed.

Valid clienttypes currently are:
- generic
- irssi
- dmdirc
- tapchat
- mirc

None of these do anything yet.