Move some settings to subclient configs (Issue #46).
This may behave oddly, needs a good chunk of testing still!

Old CWL (Issue #72) and Highlight (Issue #73) settings should be migrated automatically.

Backbuffer storage is (Issue #74) is set to the maximum of any subclient, and then shortened during sending.

TODO: In-Memory backbuffer length is only updated when the global value is changed, not sub-client values.