Rename, tidy and test CommandOutput.
I think 'CommandOutputBuffer' better reflects what the class is
actually doing. Also, rename sendBotMessage to make it clear it's
being added to the buffer rather than being *sent*.

Provide methods to directly filter the messages in a buffer,
instead of getting all the messages and replacing them (which
creates an uncessary copy, and presents a race condition).

Finally, add tests.