Add AuthProviders and make things use them.
AuthList now pokes the AuthProviders before adding/editing things in the list.
Account.checkAuthentication() now tests against the authlist.

Currently the only supported provider is "CLIENTCERT", to get started:

  /dfbnc authlist <--global|subclientname> add CLIENTCERT [fingerprint]

If fingerprint is not specified then the fingerprint of the current connection will be used if there is one.

There's some tidying up to be done on this, but it seems to work as an initial implementation of the idea.

Issue #114