Move authentication logic into its own class.
This introduces an Authenticator which is responsible for
managing the information required to auth, parsing usernames,
and dealing with passwords. The Authenticator is *nearly*
unit testable, but needs its dependency on UserSocket abstracted.

The only user-visible change is that a plain username may now
be passed in a 'PASS' command if the account has an authlist
entry that is authenticating the connection. If there is no
valid authlist entry then a plain PASS is treated as just a password.